# api/routes/hub.py
import asyncio
import logging
from pathlib import Path
from uuid import UUID

from fastapi import APIRouter, Depends, HTTPException, Request
from fastapi.responses import StreamingResponse
from pydantic import BaseModel

from api.database import get_pool
from api.services.runner import runner

logger = logging.getLogger(__name__)
router = APIRouter()


def _requires_github_token(repo_url: str) -> bool:
    candidate = Path(repo_url).expanduser()
    return not (candidate.exists() and candidate.is_dir())

async def get_current_user(request: Request) -> UUID:
    """Extract authenticated internal user UUID from the AuthMiddleware state."""
    # AuthMiddleware verifies JWT and sets workos_user_id on request.state.user_id
    workos_id = getattr(request.state, "user_id", None)
    if not workos_id:
        raise HTTPException(status_code=401, detail="Unauthorized")

    pool = get_pool()
    internal_id = await pool.fetchval("SELECT id FROM users WHERE workos_user_id = $1", workos_id)
    if not internal_id:
        raise HTTPException(status_code=401, detail="User record not found")

    return internal_id

class CreateRunRequest(BaseModel):
    repo_url: str
    prompt: str


def _create_pull_request(
    repo_url: str,
    github_token: str,
    run_id: UUID,
    summary: str,
    head_branch: str,
    base_branch: str | None,
) -> str:
    import github

    from api.routes.github import _repo_name_from_url

    gh = github.Github(github_token)
    repo = gh.get_repo(_repo_name_from_url(repo_url))
    pr = repo.create_pull(
        title=f"Autogenerated fix for Run {run_id}",
        body=f"This PR was generated by the Antigravity Hub.\n\n```\n{summary}\n```",
        head=head_branch,
        base=base_branch or getattr(repo, "default_branch", "main"),
    )
    return pr.html_url


@router.post("")
async def create_run(request: CreateRunRequest, user_id: str = Depends(get_current_user)):
    """P1-09: Dispatch a background autonomous run and save tracking state."""
    pool = get_pool()

    if _requires_github_token(request.repo_url):
        token = await pool.fetchval(
            "SELECT github_token FROM users WHERE id = $1",
            user_id
        )
        if not token:
            raise HTTPException(
                status_code=403,
                detail="You must connect a GitHub account before dispatching autonomous runs."
            )

    # Insert pending run
    run_id = await pool.fetchval(
        """
        INSERT INTO hub_runs (user_id, repo_url, task_prompt, status)
        VALUES ($1, $2, $3, 'queued') RETURNING id
        """,
        user_id, request.repo_url, request.prompt
    )

    # Spawn background task
    task = asyncio.create_task(runner.start_run(run_id, user_id, request.repo_url, request.prompt))

    # L6: Log background task exceptions
    def handle_task_result(t):
        try:
            t.result()
        except asyncio.CancelledError:
            pass
        except Exception as e:
            logger.error(f"Background hub run {run_id} failed abruptly: {e}")

    task.add_done_callback(handle_task_result)

    return {"id": run_id, "status": "queued"}


@router.get("/{run_id}/logs")
async def stream_logs(run_id: UUID, user_id: str = Depends(get_current_user)):
    """P1-09: Stream logs to frontend UI using Server-Sent Events."""
    pool = get_pool()
    # Ensure they own this run
    owner = await pool.fetchval("SELECT user_id FROM hub_runs WHERE id = $1", run_id)
    if not owner or str(owner) != str(user_id):
        raise HTTPException(status_code=404, detail="Run not found.")

    return StreamingResponse(
        runner.get_log_iterator(str(run_id)),
        media_type="text/event-stream"
    )

@router.post("/{run_id}/cancel")
async def cancel_run(run_id: UUID, user_id: str = Depends(get_current_user)):
    """P1-09: Eagerly terminate an active process."""
    pool = get_pool()
    owner = await pool.fetchval("SELECT user_id FROM hub_runs WHERE id = $1", run_id)
    if not owner or str(owner) != str(user_id):
        raise HTTPException(status_code=404, detail="Run not found.")

    cancelled = await runner.cancel_run(run_id)
    if not cancelled:
        # Check if it was already cancelled or completed
        status = await pool.fetchval("SELECT status FROM hub_runs WHERE id = $1", run_id)
        if status in ("cancelled", "completed", "failed"):
            return {"status": status, "message": "Run already inactive."}
        raise HTTPException(status_code=400, detail="Failed to cancel active run.")

    return {"status": "cancelled"}

@router.post("/{run_id}/pr")
async def create_pr(run_id: UUID, user_id: str = Depends(get_current_user)):
    """Phase 1: Verified PR Mode — block PR creation unless proof bundle is verified."""
    pool = get_pool()
    run = await pool.fetchrow(
        """
        SELECT user_id, repo_url, status, workspace_path, head_branch, base_branch
        FROM hub_runs
        WHERE id = $1
        """,
        run_id,
    )
    if not run or str(run["user_id"]) != str(user_id):
        raise HTTPException(status_code=404, detail="Run not found.")

    if run["status"] != "completed":
        raise HTTPException(status_code=400, detail="Cannot PR a run that isn't successfully completed.")

    if not run["workspace_path"] or not run["head_branch"]:
        raise HTTPException(status_code=409, detail="Run workspace metadata is missing. Re-run the Hub task.")

    # ── Phase 1 Gate: Run proof checks and validate ──────────────────
    from pathlib import Path

    from src.core.proof import ProofRunner, ProofValidator

    project_root = Path(run["workspace_path"])
    proof_runner = ProofRunner(project_root)
    bundle = proof_runner.run_all(run_id=str(run_id))

    allowed, reason = ProofValidator.validate(bundle)
    if not allowed:
        raise HTTPException(
            status_code=403,
            detail=f"PR blocked by Verified PR Mode: {reason}",
        )
    # ─────────────────────────────────────────────────────────────────

    encrypted_token = await pool.fetchval(
        "SELECT github_token FROM users WHERE id = $1",
        user_id
    )
    if not encrypted_token:
        raise HTTPException(status_code=403, detail="GitHub connection missing.")

    try:
        from api.routes.github import _decrypt_token

        gh_token = _decrypt_token(encrypted_token)
        pr_url = await asyncio.to_thread(
            _create_pull_request,
            run["repo_url"],
            gh_token,
            run_id,
            bundle.summary,
            run["head_branch"],
            run["base_branch"],
        )

        await pool.execute("UPDATE hub_runs SET pr_url = $1 WHERE id = $2", pr_url, run_id)
        runner._cleanup_workspace(str(run_id))
        return {"pr_url": pr_url, "proof": bundle.to_dict()}

    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Failed to create PR: {e}")
        raise HTTPException(status_code=500, detail="GitHub PR creation failed. Check that the bot pushed a branch cleanly.")


@router.get("/{run_id}/proof")
async def get_proof(run_id: UUID, user_id: str = Depends(get_current_user)):
    """Phase 1: Return the current proof status for a run."""
    pool = get_pool()
    run = await pool.fetchrow("SELECT user_id, workspace_path FROM hub_runs WHERE id = $1", run_id)
    if not run or str(run["user_id"]) != str(user_id):
        raise HTTPException(status_code=404, detail="Run not found.")

    from pathlib import Path

    from src.core.proof import ProofRunner

    if not run["workspace_path"]:
        raise HTTPException(status_code=409, detail="Run workspace metadata is missing. Re-run the Hub task.")

    project_root = Path(run["workspace_path"])
    bundle = ProofRunner(project_root).run_all(run_id=str(run_id))
    return {"verdict": bundle.verdict.value, "summary": bundle.summary, "proof": bundle.to_dict()}